third_party/blink/web_tests/virtual/xslt-enabled/fast/xsl/xslt-missing-namespace-in-xslt-expected.txt

CONSOLE ERROR: Namespace prefix foo on bar is not defined

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been deprecated by all browsers. These features will be removed from this browser soon. See https://chromestatus.com/feature/4709671889534976.
